Add navigation bar to website
This change introduces a navigation bar (Navbar) to the website.
A folder called "Styles" was also created to store the
CSS files related to the website, including a CSS file
dedicated to the Navbar component.

Navbar's current CSS is incomplete and needs to be refactored
for a more efficient organization and to achieve a
design more consistent with the rest of the site.

An image folder has also been created to store the resources
visuals needed for the website.